+++ src/flows/Termination/AdditionalDetailsForm.tsx
@@ -1,11 +1,8 @@
-import { JSONSchemaFormFields } from '@/src/components/form/JSONSchemaForm';
-import { Form } from '@/src/components/ui/form';
-import React, { useEffect } from 'react';
+import React from 'react';
 import { useTerminationContext } from './context';
 import { TerminationFormValues } from '@/src/flows/Termination/types';
-import { useForm } from 'react-hook-form';
-import { useJsonSchemasValidationFormResolver } from '@/src/components/form/yupValidationResolver';
 import { CreateOffboardingParams, OffboardingResponse } from '@/src/client';
+import { TerminationForm } from '@/src/flows/Termination/TerminationForm';
 
 type AdditionalDetailsFormProps = {
   username: string;
@@ -20,34 +17,7 @@ export function AdditionalDetailsForm({
   onSuccess,
   onError,
 }: AdditionalDetailsFormProps) {
-  const { formId, terminationBag } = useTerminationContext();
-
-  const resolver = useJsonSchemasValidationFormResolver(
-    // @ts-expect-error no matching type
-    terminationBag.handleValidation,
-  );
-
-  const form = useForm({
-    resolver,
-    defaultValues: terminationBag?.initialValues,
-    shouldUnregister: false,
-    mode: 'onBlur',
-  });
-
-  useEffect(() => {
-    const subscription = form?.watch((values) => {
-      const isAnyFieldDirty = Object.keys(values).some(
-        (key) =>
-          values[key as keyof TerminationFormValues] !==
-          terminationBag?.initialValues?.[key as keyof TerminationFormValues],
-      );
-      if (isAnyFieldDirty) {
-        terminationBag?.checkFieldUpdates(values);
-      }
-    });
-    return () => subscription?.unsubscribe();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  const { terminationBag } = useTerminationContext();
 
   const handleSubmit = async (values: TerminationFormValues) => {
     const terminationResult = await terminationBag?.onSubmit(values, onSubmit);
@@ -61,9 +31,7 @@ export function AdditionalDetailsForm({
     }
   };
 
-  const fields = terminationBag?.fields ? terminationBag.fields : [];
-
-  const updatedFields = fields.map((field) => {
+  const updatedFields = terminationBag?.fields.map((field) => {
     if (field.name === 'acknowledge_termination_procedure') {
       return {
         ...field,
@@ -73,15 +41,5 @@ export function AdditionalDetailsForm({
     return field;
   });
 
-  return (
-    <Form {...form}>
-      <form
-        id={formId}
-        onSubmit={form.handleSubmit(handleSubmit)}
-        className="space-y-4 RemoteFlows__TerminationForm"
-      >
-        <JSONSchemaFormFields fields={updatedFields} />
-      </form>
-    </Form>
-  );
+  return <TerminationForm fields={updatedFields} onSubmit={handleSubmit} />;
 }
