+++ src/flows/Termination/EmployeeComunicationForm.tsx
@@ -1,10 +1,7 @@
-import { JSONSchemaFormFields } from '@/src/components/form/JSONSchemaForm';
-import { Form } from '@/src/components/ui/form';
-import React, { useEffect } from 'react';
+import React from 'react';
 import { useTerminationContext } from './context';
 import { TerminationFormValues } from '@/src/flows/Termination/types';
-import { useForm } from 'react-hook-form';
-import { useJsonSchemasValidationFormResolver } from '@/src/components/form/yupValidationResolver';
+import { TerminationForm } from '@/src/flows/Termination/TerminationForm';
 
 type EmployeeComunicationProps = {
   onSubmit?: (payload: TerminationFormValues) => Promise<void>;
@@ -13,35 +10,7 @@ type EmployeeComunicationProps = {
 export function EmployeeCommunicationForm({
   onSubmit,
 }: EmployeeComunicationProps) {
-  const { formId, terminationBag } = useTerminationContext();
-
-  const resolver = useJsonSchemasValidationFormResolver(
-    // @ts-expect-error no matching type
-    terminationBag.handleValidation,
-  );
-
-  const form = useForm({
-    resolver,
-    defaultValues: terminationBag?.initialValues,
-    shouldUnregister: false,
-    mode: 'onBlur',
-  });
-
-  useEffect(() => {
-    const subscription = form?.watch((values) => {
-      const isAnyFieldDirty = Object.keys(values).some(
-        (key) =>
-          values[key as keyof TerminationFormValues] !==
-          terminationBag?.initialValues?.[key as keyof TerminationFormValues],
-      );
-      if (isAnyFieldDirty) {
-        terminationBag?.checkFieldUpdates(values);
-      }
-    });
-    return () => subscription?.unsubscribe();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
+  const { terminationBag } = useTerminationContext();
   const handleSubmit = async (values: TerminationFormValues) => {
     await onSubmit?.(
       terminationBag?.parseFormValues(values) as TerminationFormValues,
@@ -49,17 +18,5 @@ export function EmployeeCommunicationForm({
     terminationBag?.next();
   };
 
-  const fields = terminationBag?.fields ? terminationBag.fields : [];
-
-  return (
-    <Form {...form}>
-      <form
-        id={formId}
-        onSubmit={form.handleSubmit(handleSubmit)}
-        className="space-y-4 RemoteFlows__TerminationForm"
-      >
-        <JSONSchemaFormFields fields={fields} />
-      </form>
-    </Form>
-  );
+  return <TerminationForm onSubmit={handleSubmit} />;
 }
