+++ src/flows/Termination/index.ts
@@ -1,5 +1,3 @@
 export { TerminationFlow } from './TerminationFlow';
 export type { TerminationRenderProps } from './TerminationFlow';
-export { TerminationForm } from './TerminationForm';
-export { TerminationSubmit } from './TerminationSubmit';
 export type { TerminationFormValues } from './types';
